FIX: isl29028: Exposes 'polling_ignore_enabled' to userspace via sysfs.
Due to faulty factory calibration in the Desire HD, the proximity
sensor needs to be recalibrated. HTC's 'polling' workaround works
for probably 90% of the cases. This patch is for the remaining 10%
that still are not satisfied.

Polling can be disabled by writing 0 to the ps_polling_ignore.
To re-enable, write 1.

Polling, if not disabled, gets activated everytime the sensor is
enabled, and thus overwrites the user-configured calibration.
Interestingly, the code does restore the user setting when the
sensor gets disabled.