msm: kgsl: Use ringbuffer WPTR drop prevention workaround only for a20x GPU  - modified for DHD by LorD ClockaN
Sometimes, the writes to ringbuffer WPTR register are dropped. This will
make the GPU to go to idle before completing all the commands that it is
submitted with. This workaround prevents the writes from being dropped.

The workaround added to prevent the writes to ringbuffer WPTR from
being dropped does not work on a22x GPU. So, restrict this workaround
only for a20x GPU.

CRs-fixed: 305417

Change-Id: I8d15cf3c7b2356e1c8153cc7e8cffe158321c111